LLM/Gemini/requirements.txt
google-generativeai>=0.3.0
google-genai>=0.2.0
Pillow>=10.0.0
requests>=2.28.0

